refactor: moved 'AuthConnection' enum to shared constants
