+++ shared/constants/oauth.ts
@@ -1,0 +1,10 @@
+export enum AuthConnection {
+  /**
+   * Google
+   */
+  Google = 'google',
+  /**
+   * Apple
+   */
+  Apple = 'apple',
+}
